Move payload builders into one module
