--- src/functions/buildPayload.js
@@ -1,150 +1,0 @@
-/**
- * (c) Copyright Reserved EVRYTHNG Limited 2018.
- * All rights reserved. Use of this material is subject to license.
- */
-
-const evrythngSwagger = require('evrythng-swagger');
-const jsonSchemaParser = require('json-schema-ref-parser');
-const prompt = require('../modules/prompt');
-const indent = require('./indent');
-const logger = require('../modules/logger');
-
-// Some properties are writable, just not at create time
-const NON_CREATE_PROPERTIES = [
-  'scopes', 'properties', 'fn', 'startsAt', 'endsAt', 'location',
-];
-
-// Values that are always the same value, and are not in the spec
-const HARDCODE_MAP = {
-  socialNetworks: {},
-};
-
-const SPECIAL_BUILDERS = {
-  task: require('./buildTask'),
-};
-
-let spec;
-
-const getKeyName = async () => prompt.getValue(indent('key', 2));
-
-const buildCustomObject = async () => {
-  const result = {};
-
-  let key = await getKeyName();
-  while (key) {
-    const value = await prompt.getValue(indent('value', 2));
-    result[key] = value;
-
-    key = await getKeyName();
-  }
-
-  return result;
-};
-
-const buildDefObject = async (opts) => {
-  const { key, target, index, total, propertyDef } = opts;
-
-  // User-definable fields are not documented, present free-form
-  if (['customFields', 'identifiers'].includes(key)) {
-    logger.info(`${index + 1}/${total}: ${key} (object, leave 'key' blank to finish)`);
-    target[key] = await buildCustomObject();
-    return;
-  }
-
-  // Recurse
-  target[key] = {};
-  await buildObject(target[key], propertyDef.properties, key);
-};
-
-const buildDefProperty = async (opts) => {
-  const { index, total, target, key, propertyDef } = opts;
-
-  // Fields with only one allowed value
-  if (HARDCODE_MAP[key]) {
-    target[key] = HARDCODE_MAP[key];
-    logger.info(`${index + 1}/${total}: ${key} is always ${JSON.stringify(HARDCODE_MAP[key])}`);
-    return;
-  }
-
-  // Build sub-object
-  // TODO handle sub-object properties
-  if (propertyDef.type === 'object') {
-    await buildDefObject({ key, target, index, total, propertyDef });
-    return;
-  }
-
-  // Determine type hint string
-  let typeStr = propertyDef.type;
-  if (propertyDef.type === 'array') {
-    typeStr = `comma-separated list of ${propertyDef.items.type}`;
-  }
-  if (propertyDef.enum) {
-    typeStr = `${propertyDef.type}, one of '${propertyDef.enum.join('\', \'')}'`;
-  }
-
-  // Get a simple value
-  const input = await prompt.getValue(`${index + 1}/${total}: ${key} (${typeStr})`);
-  if (!input) {
-    return;
-  }
-
-  // Handle input arrays as comma-separated values
-  // TODO handle arrays of non-strings?
-  if (propertyDef.type === 'array') {
-    if (propertyDef.items.type === 'string') {
-      target[key] = input.split(',');
-      return;
-    }
-
-    if (['integer', 'number'].includes(propertyDef.items.type)) {
-      target[key] = input.split(',').map(parseInt);
-      return;
-    }
-  }
-
-  // Simple value
-  target[key] = input;
-};
-
-const filteredPropertyKeys = properties => Object.keys(properties)
-  .filter(item => !properties[item].readOnly)
-  .filter(item => !NON_CREATE_PROPERTIES.includes(item));
-
-const buildObject = async (properties, name) => {
-  const propertyKeys = filteredPropertyKeys(properties);
-  const context = name ? `of ${name} ` : '';
-  logger.info(`Provide values for each field ${context}(or leave blank to skip):\n`);
-
-  const payload = {};
-  for (let index = 0; index < propertyKeys.length; index += 1) {
-    const key = propertyKeys[index];
-    await buildDefProperty({
-      key,
-      index,
-      total: propertyKeys.length,
-      target: payload,
-      propertyDef: properties[key],
-    });
-  }
-
-  return payload;
-};
-
-module.exports = async (defName) => {
-  // Special builders
-  if (SPECIAL_BUILDERS[defName]) {
-    return SPECIAL_BUILDERS[defName]();
-  }
-
-  spec = await jsonSchemaParser.dereference(evrythngSwagger);
-  if (!spec.definitions[defName]) {
-    throw new Error(`\ndefName ${defName} not found in spec!`);
-  }
-
-  logger.info();
-  const { properties } = spec.definitions[defName];
-  const payload = await buildObject(properties);
-
-  logger.info();
-  return payload;
-};
